move ethernet & uart code to separate files
